gfx/layers/apz/test/mochitest/test_tap.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Sanity panning test</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

SimpleTest.waitForExplicitFinish();

// this page just serially loads each one of the following test helper pages in
// a new window and waits for it to call testDone()
var tests = [
  {'file': 'helper_tap.html'},
  // For the following two tests, disable displayport suppression to make sure it
  // doesn't interfere with the test by scheduling paints non-deterministically.
  {'file': 'helper_scrollto_tap.html?true', 'prefs': [["apz.paint_skipping.enabled", true]], 'dp_suppression': false},
  {'file': 'helper_scrollto_tap.html?false', 'prefs': [["apz.paint_skipping.enabled", false]], 'dp_suppression': false},
  // For the long-tap test, reduce the content response timeout because the touchstart
  // event doesn't get processed (because of the event listener) until this expires.
  // Once we support passive event listeners, we can use that instead and stop mucking
  // with the timeout.
  {'file': 'helper_long_tap.html'}
];

var testIndex = -1;
var w = null;

function testDone() {
  var test = tests[testIndex];
  if (w) {
    if (typeof test.dp_suppression != 'undefined') {
      // We modified the suppression when starting the test, so now undo that.
      SpecialPowers.getDOMWindowUtils(window).respectDisplayPortSuppression(!test.dp_suppression);
    }
    if (!!test.prefs) {
      // We pushed some prefs for this test, pop them, and re-invoke
      // testDone() after that's been processed
      SpecialPowers.popPrefEnv(function() {
        w.close();
        w = null;
        testDone();
      });
      return;
    }

    w.close();
  }

  testIndex++;
  if (testIndex >= tests.length) {
    SimpleTest.finish();
    return;
  }

  test = tests[testIndex];
  if (typeof test.dp_suppression != 'undefined') {
    // Normally during a test, the displayport will get suppressed during page
    // load, and unsuppressed at a non-deterministic time during the test. The
    // unsuppression can trigger a repaint which interferes with the test, so
    // to avoid that we can force the displayport to be unsuppressed for the
    // entire test which is more deterministic.
    SpecialPowers.getDOMWindowUtils(window).respectDisplayPortSuppression(test.dp_suppression);
  }
  if (!!test.prefs) {
    // Got some prefs for this subtest, push them
    SpecialPowers.pushPrefEnv({"set": test.prefs}, function() {
      w = window.open(test.file, "_blank");
    });
  } else {
    w = window.open(test.file, "_blank");
  }
}

window.onload = function() {
  if (!SpecialPowers.getDOMWindowUtils(window).asyncPanZoomEnabled) {
    ok(true, "APZ is not enabled, this test is not relevant, sorry!\n");
    SimpleTest.finish();
    return;
  }
  testDone();
};

  </script>
</head>
<body>
</body>
</html>
